Wait for showifs and such to avoid more race conditions
This really isn't sufficient, but I'm not sure what would be.
Sometimes even on load things don't appear for a moment or two and
then move stuff when they do appear. Maybe wait 10ms to see if
stuff looks different than it did before and then, if not, carry on
and, if so, wait a little longer?